+++ src/extension/ios/plistBuddy.ts
@@ -7,14 +7,23 @@ import * as glob from "glob";
 import * as fs from "fs";
 import * as semver from "semver";
 
-import {Node} from "../../common/node/node";
-import {ChildProcess} from "../../common/node/childProcess";
+import { Node } from "../../common/node/node";
+import { ChildProcess } from "../../common/node/childProcess";
 import { ErrorHelper } from "../../common/error/errorHelper";
 import { InternalErrorCode } from "../../common/error/internalErrorCode";
 import { ProjectVersionHelper } from "../../common/projectVersionHelper";
+import { getFileNameWithoutExtension } from "../../common/utils";
+
+export interface ConfigurationData {
+    fullProductName: string;
+    configurationFolder: string;
+}
 
 export class PlistBuddy {
     private static plistBuddyExecutable = "/usr/libexec/PlistBuddy";
+
+    private readonly TARGET_BUILD_DIR_SEARCH_KEY = "TARGET_BUILD_DIR";
+    private readonly FULL_PRODUCT_NAME_SEARCH_KEY = "FULL_PRODUCT_NAME";
 
     private nodeChildProcess: ChildProcess;
 
@@ -37,14 +46,38 @@ export class PlistBuddy {
             } else {
                 productsFolder = path.join(iosProjectRoot, "build", "Build", "Products");
             }
-            const configurationFolder = path.join(productsFolder, `${configuration}${simulator ? "-iphonesimulator" : "-iphoneos"}`);
+            const sdkType = simulator ? "iphonesimulator" : "iphoneos";
+            let configurationFolder = path.join(productsFolder, `${configuration}-${sdkType}`);
             let executable = "";
             if (productName) {
                 executable = `${productName}.app`;
+                if (!fs.existsSync(path.join(configurationFolder, executable))) {
+                    const configurationData = this.getConfigurationData(
+                        projectRoot,
+                        rnVersions.reactNativeVersion,
+                        iosProjectRoot,
+                        configuration,
+                        scheme,
+                        sdkType,
+                        configurationFolder
+                    );
+                    configurationFolder = configurationData.configurationFolder;
+                }
             } else {
                 const executableList = this.findExecutable(configurationFolder);
                 if (!executableList.length) {
-                    throw ErrorHelper.getInternalError(InternalErrorCode.IOSCouldNotFoundExecutableInFolder, configurationFolder);
+                    const configurationData = this.getConfigurationData(
+                        projectRoot,
+                        rnVersions.reactNativeVersion,
+                        iosProjectRoot,
+                        configuration,
+                        scheme,
+                        sdkType,
+                        configurationFolder
+                    );
+
+                    configurationFolder = configurationData.configurationFolder;
+                    executableList.push(configurationData.fullProductName);
                 } else if (executableList.length > 1) {
                     throw ErrorHelper.getInternalError(InternalErrorCode.IOSFoundMoreThanOneExecutablesCleanupBuildFolder, configurationFolder);
                 }
@@ -80,7 +113,54 @@ export class PlistBuddy {
         return this.invokePlistBuddy(`Print ${property}`, plistFile);
     }
 
+    public getBuildPathAndProductName(
+        iosProjectRoot: string,
+        projectWorkspaceConfigName: string,
+        configuration: string,
+        scheme: string,
+        sdkType: string
+    ): ConfigurationData {
+        const buildSettings = this.nodeChildProcess.execFileSync(
+            "xcodebuild",
+            [
+                "-workspace",
+                projectWorkspaceConfigName,
+                "-scheme",
+                scheme,
+                "-sdk",
+                sdkType,
+                "-configuration",
+                configuration,
+                "-showBuildSettings",
+            ],
+            {
+                encoding: "utf8",
+                cwd: iosProjectRoot,
+            }
+        );
+
+        const targetBuildDir = this.fetchParameterFromBuildSettings(<string>buildSettings, this.TARGET_BUILD_DIR_SEARCH_KEY);
+        const fullProductName = this.fetchParameterFromBuildSettings(<string>buildSettings, this.FULL_PRODUCT_NAME_SEARCH_KEY);
+
+        if (!targetBuildDir) {
+            throw new Error("Failed to get the target build directory.");
+        }
+        if (!fullProductName) {
+            throw new Error("Failed to get full product name.");
+        }
+
+        return {
+            fullProductName,
+            configurationFolder: targetBuildDir,
+        };
+    }
+
     public getInferredScheme(iosProjectRoot: string, projectRoot: string, rnVersion: string) {
+        const projectWorkspaceConfigName = this.getProjectWorkspaceConfigName(iosProjectRoot, projectRoot, rnVersion);
+        return getFileNameWithoutExtension(projectWorkspaceConfigName);
+    }
+
+    public getProjectWorkspaceConfigName(iosProjectRoot: string, projectRoot: string, rnVersion: string): string {
         // Portion of code was taken from https://github.com/react-native-community/cli/blob/master/packages/platform-ios/src/commands/runIOS/index.js
         // and modified a little bit
         /**
@@ -106,11 +186,41 @@ export class PlistBuddy {
             );
         }
 
-        const inferredSchemeName = path.basename(
-            xcodeProject.name,
-            path.extname(xcodeProject.name)
+        return xcodeProject.name;
+    }
+
+    public getConfigurationData(
+        projectRoot: string,
+        reactNativeVersion: string,
+        iosProjectRoot: string,
+        configuration: string,
+        scheme: string | undefined,
+        sdkType: string,
+        oldConfigurationFolder: string
+    ): ConfigurationData {
+        if (!scheme) {
+            throw ErrorHelper.getInternalError(InternalErrorCode.IOSCouldNotFoundExecutableInFolder, oldConfigurationFolder);
+        }
+        const projectWorkspaceConfigName = this.getProjectWorkspaceConfigName(iosProjectRoot, projectRoot, reactNativeVersion);
+        return this.getBuildPathAndProductName(
+            iosProjectRoot,
+            projectWorkspaceConfigName,
+            configuration,
+            scheme,
+            sdkType
         );
-        return inferredSchemeName;
+    }
+
+    /**
+     * @param {string} buildSettings
+     * @param {string} parameterName
+     * @returns {string | null}
+     */
+    public fetchParameterFromBuildSettings(buildSettings: string, parameterName: string) {
+        const targetBuildMatch = new RegExp(`${parameterName} = (.+)$`, "m").exec(buildSettings);
+        return targetBuildMatch && targetBuildMatch[1]
+            ? targetBuildMatch[1].trim()
+            : null;
     }
 
     private findExecutable(folder: string): string[] {