+++ test/smoke/automation/src/debug.ts
@@ -7,11 +7,10 @@ import { Code, findElement } from "./code";
 import { Editors } from "./editors";
 import { Editor } from "./editor";
 import { IElement } from "../src/driver";
-import { QuickOpen } from ".";
 
 const VIEWLET = "div[id=\"workbench.view.debug\"]";
-const DEBUG_VIEW = `${VIEWLET} .debug-view-content`;
-const CONFIGURE = `div[id="workbench.parts.sidebar"] .actions-container .configure`;
+const DEBUG_VIEW = `${VIEWLET}`;
+const CONFIGURE = `div[id="workbench.parts.sidebar"] .actions-container .codicon-gear`;
 const STOP = `.debug-toolbar .action-label[title*="Stop"]`;
 const STEP_OVER = `.debug-toolbar .action-label[title*="Step Over"]`;
 const STEP_IN = `.debug-toolbar .action-label[title*="Step Into"]`;
@@ -19,7 +18,8 @@ const STEP_OUT = `.debug-toolbar .action-label[title*="Step Out"]`;
 const CONTINUE = `.debug-toolbar .action-label[title*="Continue"]`;
 const DISCONNECT = `.debug-toolbar .action-label[title*="Disconnect"]`;
 const GLYPH_AREA = ".margin-view-overlays>:nth-child";
-const BREAKPOINT_GLYPH = ".debug-breakpoint";
+const BREAKPOINT_GLYPH = ".codicon-debug-breakpoint";
+const PAUSE = `.debug-toolbar .action-label[title*="Pause"]`;
 const DEBUG_STATUS_BAR = `.statusbar.debugging`;
 const NOT_DEBUG_STATUS_BAR = `.statusbar:not(debugging)`;
 const TOOLBAR_HIDDEN = `.debug-toolbar[aria-hidden="true"]`;
@@ -50,7 +50,7 @@ function toStackFrame(element: IElement): IStackFrame {
 
 export class Debug extends Viewlet {
 
-    constructor(code: Code, private commands: Commands, private editors: Editors, private editor: Editor, private quickopen: QuickOpen) {
+    constructor(code: Code, private commands: Commands, private editors: Editors, private editor: Editor) {
         super(code);
     }
 
@@ -75,8 +75,16 @@ export class Debug extends Viewlet {
         await this.code.waitForElement(BREAKPOINT_GLYPH);
     }
 
-    public async startDebugging(): Promise<void> {
+    public async startDebugging(): Promise<number> {
         await this.code.dispatchKeybinding("f5");
+        await this.code.waitForElement(PAUSE);
+        await this.code.waitForElement(DEBUG_STATUS_BAR);
+        const portPrefix = "Port: ";
+
+        const output = await this.waitForOutput(output => output.some(line => line.indexOf(portPrefix) >= 0));
+        const lastOutput = output.filter(line => line.indexOf(portPrefix) >= 0)[0];
+
+        return lastOutput ? parseInt(lastOutput.substr(portPrefix.length)) : 3000;
     }
 
     public async waitForDebuggingToStart(): Promise<void> {
@@ -85,11 +93,6 @@ export class Debug extends Viewlet {
 
     public async areStackFramesExist(): Promise<any> {
         return await this.code.waitForElement(STACK_FRAME);
-    }
-
-    public async runDebugScenario(debugOption: string): Promise<any> {
-        await this.quickopen.openQuickOpen();
-        await this.quickopen.submit(`debug ${debugOption}`);
     }
 
     public async stepOver(): Promise<any> {