+++ test/smoke/package/src/helpers/smokeTestsConstants.ts
@@ -16,7 +16,7 @@ export class SmokeTestsConstants {
     // Timeout for driver to wait for UI elements response to interaction (in seconds)
     public static elementResponseTimeout = 250;
     // Timeout for enabling Remote JS Debugging while testing RN app
-    public static enableRemoteJSTimeout = 50 * 1000;
+    public static enableRemoteJSTimeout = 120 * 1000;
     // Timeout for Android app to build and to install
     public static androidAppBuildAndInstallTimeout = 300 * 1000;
     // Timeout for iOS app to build and to install